+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   - Block Purchase button (NEW): prevents players from buying the item
   - Block Sale button: prevents players from selling the item
 - The Store Randomizer allows the GM to dynamically generate the store's inventory. Instead of manually hiding or pricing items, you can set parameters for each category and let the module "roll" the stock for you.
-- 
+- Items with description will have an tooltip 
 
 # 0.1.1
 - Performance fix: the store loads 3,04 faster.